backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# PostgreSQL drivers (for pgAdmin 4 compatibility)
asyncpg==0.29.0
psycopg2-binary==2.9.7

# SQLite driver (for development)
aiosqlite==0.19.0

# AI/ML Services
openai==1.3.8
spacy==3.7.2
transformers==4.35.2
torch>=2.0.0
librosa==0.10.1
soundfile==0.12.1

# Google Cloud Speech (Alternative to Whisper)
google-cloud-speech==2.22.0
google-auth==2.23.4

# Voice Processing
pydub==0.25.1
wave==0.0.2
pyaudio==0.2.11

# HTTP Client for external APIs
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data Models and Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Security and Authentication
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
python-slugify==8.0.1
email-validator==2.1.0

# Date and Time
python-dateutil==2.8.2

# File Processing
Pillow==10.1.0
python-magic==0.4.27
aiofiles==23.2.1

# Logging and Monitoring
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0

# WebSocket support
websockets==12.0

# Rate limiting
slowapi==0.1.9

# Background tasks
celery==5.3.4
redis==5.0.1

# Optional: For production deployment
gunicorn==21.2.0
prometheus-fastapi-instrumentator==6.1.0 